+++ lib/pages/share_two.dart
@@ -63,6 +63,9 @@ class ShareStateTwoState extends State<ShareTwo> {
               child: Padding(
                 padding: const EdgeInsets.only(left: 16.0, right: 16.0),
                 child: TextFormField(
+                  autocorrect: false,
+                  enableSuggestions: false,
+                  keyboardType: TextInputType.emailAddress,
                   decoration: InputDecoration(
                     hintText: getString(context, "share_list"),
                     suffixIcon: Icon(